requirements.txt
click==7.1.2
comtypes==1.1.8
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
PyPDF2==1.26.0
pypiwin32==223
python-docx==0.8.10
pyttsx3==2.90
pywin32==300
Werkzeug==1.0.1
